started evaluations

+++ edge/prepopulate.py
@@ -14,29 +14,30 @@ def populate_database():
     Station.objects.all().delete()
     Upload.objects.all().delete()
 
-    shutil.rmtree('media')
+    if os.path.exists("media"):
+        shutil.rmtree("media")
 
-    company, _ = Company.objects.get_or_create(
-        companyname="exxon"
-    )
-    sample_image_path = os.path.join('prepopulate_data', 'prepopulate_sign.jpg')
-    image = Image()
-    image.imagefield.save('prepopulate_sign.jpg', File(open(sample_image_path, 'rb')))
-
-    station, _ = Station.objects.get_or_create(
-        company = company,
-        latitude = 14.10618,
-        longitude = -44.78551
-    )
-
-    Upload.objects.get_or_create(
-        latitude = 14.10618,
-        longitude = -44.78551,
-        timestamp = datetime.datetime.now(),
-        station = station,
-        price = "1.99",
-        image = image
-    )
+    # company, _ = Company.objects.get_or_create(
+    #     companyname="exxon"
+    # )
+    # sample_image_path = os.path.join('prepopulate_data', 'prepopulate_sign.jpg')
+    # image = Image()
+    # image.imagefield.save('prepopulate_sign.jpg', File(open(sample_image_path, 'rb')))
+    #
+    # station, _ = Station.objects.get_or_create(
+    #     company = company,
+    #     latitude = 14.10618,
+    #     longitude = -44.78551
+    # )
+    #
+    # Upload.objects.get_or_create(
+    #     latitude = 14.10618,
+    #     longitude = -44.78551,
+    #     timestamp = datetime.datetime.now(),
+    #     station = station,
+    #     price = "1.99",
+    #     image = image
+    # )
 
 
 # Start execution here!
